Reject unrecognized enum constants with an exception.
Also promote enum adapter to StandardJsonAdapters.java.
