+++ moshi/src/main/java/com/squareup/moshi/Moshi.java
@@ -34,7 +34,6 @@ public final class Moshi {
     List<JsonAdapter.Factory> factories = new ArrayList<>();
     factories.addAll(builder.factories);
     factories.add(StandardJsonAdapters.FACTORY);
-    factories.add(EnumJsonAdapter.FACTORY);
     factories.add(CollectionJsonAdapter.FACTORY);
     factories.add(MapJsonAdapter.FACTORY);
     factories.add(ArrayJsonAdapter.FACTORY);
